Move keymanager e2e test to cli
