+++ packages/lodestar/test/utils/node/validator.ts
@@ -1,112 +1,10 @@
-import fs from "node:fs";
-import path from "node:path";
-import tmp, {DirResult, FileResult} from "tmp";
+import tmp from "tmp";
 import {LevelDbController} from "@chainsafe/lodestar-db";
 import {interopSecretKey} from "@chainsafe/lodestar-beacon-state-transition";
-import {
-  SlashingProtection,
-  Validator,
-  Signer,
-  SignerType,
-  ISlashingProtection,
-  SignerLocal,
-} from "@chainsafe/lodestar-validator";
+import {SlashingProtection, Validator, Signer, SignerType} from "@chainsafe/lodestar-validator";
 import type {SecretKey} from "@chainsafe/bls/types";
-import {KEYSTORE_IMPORTED_PREFIX} from "@chainsafe/lodestar-keymanager-server";
 import {BeaconNode} from "../../../src/node/index.js";
 import {testLogger, TestLoggerOpts} from "../logger.js";
-import {getLocalSecretKeys} from "../../../../cli/src/cmds/validator/keys.js";
-import {IValidatorCliArgs} from "../../../../cli/src/cmds/validator/options.js";
-import {IGlobalArgs} from "../../../../cli/src/options/index.js";
-
-export async function getAndInitValidatorsWithKeystore({
-  node,
-  keystoreContent,
-  keystorePubKey,
-  useRestApi,
-  testLoggerOpts,
-}: {
-  node: BeaconNode;
-  keystoreContent: string;
-  keystorePubKey: string;
-  useRestApi?: boolean;
-  testLoggerOpts?: TestLoggerOpts;
-}): Promise<{
-  validator: Validator;
-  secretKeys: SecretKey[];
-  keystoreContent: string;
-  signers: SignerLocal[];
-  slashingProtection: ISlashingProtection;
-  tempDirs: {
-    keystoreDir: DirResult;
-    passwordFile: FileResult;
-  };
-}> {
-  const keystoreDir = tmp.dirSync({unsafeCleanup: true});
-  // TODO: This hardcoded value is necessary for a keymanager test
-  const keystoreFile = path.join(`${keystoreDir.name}`, `${KEYSTORE_IMPORTED_PREFIX}_${keystorePubKey}.json`);
-
-  fs.writeFileSync(keystoreFile, keystoreContent, {encoding: "utf8", flag: "wx"});
-
-  const passwordFile = tmp.fileSync();
-  fs.writeFileSync(passwordFile.name, "test123!", {encoding: "utf8"});
-
-  const vcConfig = {
-    network: "prater",
-    importKeystoresPath: [`${keystoreDir.name}`],
-    importKeystoresPassword: `${passwordFile.name}`,
-    keymanagerEnabled: true,
-    keymanagerAuthEnabled: true,
-    keymanagerHost: "127.0.0.1",
-    keymanagerPort: 9666,
-    keymanagerCors: "*",
-  };
-
-  const logger = testLogger("Vali", testLoggerOpts);
-  const tmpDir = tmp.dirSync({unsafeCleanup: true});
-  const dbOps = {
-    config: node.config,
-    controller: new LevelDbController({name: tmpDir.name}, {logger}),
-  };
-  const slashingProtection = new SlashingProtection(dbOps);
-
-  const signers: SignerLocal[] = [];
-
-  const {secretKeys, unlockSecretKeys: _unlockSecretKeys} = await getLocalSecretKeys(
-    (vcConfig as unknown) as IValidatorCliArgs & IGlobalArgs
-  );
-  if (secretKeys.length > 0) {
-    // Log pubkeys for auditing
-    logger.info(`Decrypted ${secretKeys.length} local keystores`);
-    for (const secretKey of secretKeys) {
-      logger.info(secretKey.toPublicKey().toHex());
-      signers.push({
-        type: SignerType.Local,
-        secretKey,
-      });
-    }
-  }
-
-  const validator = await Validator.initializeFromBeaconNode({
-    dbOps,
-    api: useRestApi ? getNodeApiUrl(node) : node.api,
-    slashingProtection,
-    logger,
-    signers,
-  });
-
-  return {
-    validator,
-    secretKeys,
-    keystoreContent,
-    signers,
-    slashingProtection,
-    tempDirs: {
-      keystoreDir: keystoreDir,
-      passwordFile: passwordFile,
-    },
-  };
-}
 
 export async function getAndInitDevValidators({
   node,
@@ -148,8 +46,8 @@ export async function getAndInitDevValidators({
       ? secretKeysValidator.map(
           (secretKey): Signer => ({
             type: SignerType.Remote,
-            externalSignerUrl,
-            pubkeyHex: secretKey.toPublicKey().toHex(),
+            url: externalSignerUrl,
+            pubkey: secretKey.toPublicKey().toHex(),
           })
         )
       : secretKeysValidator.map(
@@ -179,7 +77,7 @@ export async function getAndInitDevValidators({
 }
 
 function getNodeApiUrl(node: BeaconNode): string {
-  const host = node.opts.api.rest.host || "127.0.0.1";
+  const address = node.opts.api.rest.address || "127.0.0.1";
   const port = node.opts.api.rest.port || 19596;
-  return `http://${host}:${port}`;
+  return `http://${address}:${port}`;
 }
